Add files via upload
Train seat reservation system
In a train seat reservation system, A user (customer/railway employee) will search for train in railway server. To know the availability the user, must select the details of the train so that he/she can be sure about confirming. First, he/she will see the destination and select one. Then he/she will be asked to select the preferable date and time. After selecting, the server will see if there is any train available of that time. If yes then server will ask for further details and if not, the server will deny. For the available train, user will select the train and check no. of seat which are available. Then server asks if the user is interested to go further for reservation. Then comes the booking process. To confirm the seat the user, have to login to the server as a passenger or railway employee. After verifying the information of the user by system, the user will select train details and then system will show the amount to pay. If the user wants to confirm the reservation, he/she must pay the bill otherwise the ticket will be cancelled. To pay the amount the user will enter the credit card info and verify it or cash. If the information is nor verified it will ask again to enter correct details. After verifying the user will pay the bill and confirm the reservation. After confirming if user want she/he can cancel reservation